agent/requirements.txt
strands-agents
boto3
python-dotenv
fastapi
uvicorn
gunicorn
pydantic>=2.0
fastmcp>=2.4.0
nest-asyncio
awslabs-iam-mcp-server>=1.0.6
awslabs-cloudtrail-mcp-server>=0.0.4
awslabs-aws-api-mcp-server>=0.1.0
PyJWT>=2.8.0
requests>=2.31.0